requirements.txt
Django==1.10.5
psycopg2==2.6.2
djangorestframework==3.5.4
dj-database-url==0.4.2
gunicorn==19.6.0
